Long running tests are converted to IT tests

+++ lucene/src/test/java/com/orientechnologies/lucene/test/LuceneAutomaticBackupRestoreTest.java
@@ -34,6 +34,8 @@ import com.orientechnologies.orient.server.config.OServerParameterConfiguration;
 import com.orientechnologies.orient.server.handler.OAutomaticBackup;
 import org.junit.*;
 import org.junit.rules.TemporaryFolder;
+import org.junit.runner.RunWith;
+import org.junit.runners.JUnit4;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -51,6 +53,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 /**
  * Created by Enrico Risa on 07/07/15.
  */
+@RunWith(JUnit4.class)
 public class LuceneAutomaticBackupRestoreTest {
 
   private final static String          DBNAME     = "OLuceneAutomaticBackupRestoreTest";
@@ -66,6 +69,10 @@ public class LuceneAutomaticBackupRestoreTest {
 
   @Before
   public void setUp() throws Exception {
+
+    final String os = System.getProperty("os.name").toLowerCase();
+    Assume.assumeFalse(os.contains("win"));
+
 
     server = new OServer() {
       @Override
@@ -114,10 +121,12 @@ public class LuceneAutomaticBackupRestoreTest {
 
   @After
   public void tearDown() throws Exception {
-    dropIfExists();
+    final String os = System.getProperty("os.name").toLowerCase();
+    if (!os.contains("win")) {
+      dropIfExists();
 
-    tempFolder.delete();
-
+      tempFolder.delete();
+    }
   }
 
   @Test
